- Summary: fix bug117061: autoscaling for y axis - Bugzilla Bug (s) Resolved: 117061 - Description: 1. Now use approximate actual mininum/maximum value to compute the scale. If you want a fixed min/max values, please specify explicitly in scale settings. 2. code cleanup in chart model. - Tests Description  :
- Files Edited:

- Files Added:

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

+++ chart/org.eclipse.birt.chart.engine/src/org/eclipse/birt/chart/render/AxesRenderer.java
@@ -2052,7 +2052,7 @@ public abstract class AxesRenderer extends BaseRenderer
 								&& insCA.getRight( ) < lia.getThickness( ) )
 							continue;
 
-						y = (int) ( da[j] - pwa.getSeriesThickness( ) );
+						y = ( da[j] - pwa.getSeriesThickness( ) );
 						lre = (LineRenderEvent) ( (EventObjectCache) ipr ).getEventObject( StructureSource.createPlot( p ),
 								LineRenderEvent.class );
 						lre.setLineAttributes( lia );
@@ -2071,7 +2071,7 @@ public abstract class AxesRenderer extends BaseRenderer
 							&& insCA.getRight( ) < lia.getThickness( ) )
 						continue;
 
-					y = (int) da[j];
+					y = da[j];
 					if ( pwa.getDimension( ) == IConstants.TWO_5_D )
 					{
 						y -= pwa.getSeriesThickness( );
